rename and add i2c
