+++ emBODY/eBcode/arch-arm/board/amc/application/v2/proj/amc-icc.uvprojx
@@ -854,6 +854,11 @@
               <FilePath>..\..\..\bsp\embot_hw_icc_bsp_amc.cpp</FilePath>
             </File>
             <File>
+              <FileName>embot_hw_i2c_bsp_amc.cpp</FileName>
+              <FileType>8</FileType>
+              <FilePath>..\..\..\bsp\embot_hw_i2c_bsp_amc.cpp</FilePath>
+            </File>
+            <File>
               <FileName>embot_hw_led_bsp_amc.cpp</FileName>
               <FileType>8</FileType>
               <FilePath>..\..\..\bsp\embot_hw_led_bsp_amc.cpp</FilePath>
@@ -869,6 +874,11 @@
               <FilePath>..\..\..\bsp\embot_hw_spi_bsp_amc.cpp</FilePath>
             </File>
             <File>
+              <FileName>embot_hw_specialize_bsp_amc.cpp</FileName>
+              <FileType>8</FileType>
+              <FilePath>..\..\..\bsp\embot_hw_specialize_bsp_amc.cpp</FilePath>
+            </File>
+            <File>
               <FileName>embot_hw_timer_bsp_amc.cpp</FileName>
               <FileType>8</FileType>
               <FilePath>..\..\..\bsp\embot_hw_timer_bsp_amc.cpp</FilePath>
@@ -877,11 +887,6 @@
               <FileName>embot_hw_testpoint_bsp_amc .cpp</FileName>
               <FileType>8</FileType>
               <FilePath>..\..\..\bsp\embot_hw_testpoint_bsp_amc .cpp</FilePath>
-            </File>
-            <File>
-              <FileName>embot_hw_bsp_amc.cpp</FileName>
-              <FileType>8</FileType>
-              <FilePath>..\..\..\bsp\embot_hw_bsp_amc.cpp</FilePath>
             </File>
           </Files>
         </Group>
