+++ emBODY/eBcode/arch-arm/board/amc/bsp/embot_hw_bsp_amc_config.h
@@ -39,7 +39,6 @@
 //    #define EMBOT_ENABLE_hw_button
     #define EMBOT_ENABLE_hw_can
     #define EMBOT_ENABLE_hw_flash
-//    #define EMBOT_ENABLE_hw_eth
 
     #define EMBOT_ENABLE_hw_eth
     #define EMBOT_ENABLE_hw_timer
